requirements.txt
streamlit
google-generativeai
fpdf
python-dotenv
requests
beautifulsoup4
reportlab